+++ audiobook/src/main/kotlin/de/ph1b/audiobook/fragment/SettingsFragment.kt
@@ -59,7 +59,6 @@ class SettingsFragment : PreferenceFragment(), DonationDialogFragment.OnDonation
         super.onDestroy()
 
         iabHelper.dispose()
-        App.leakWatch(this)
     }
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {