+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Next Version
+* Added a new cover picker
 * Fixed a crash when theming a dialog on Android 4.0.4
 * Fixes to the internal player
 * Fallback to the Android player when there is a bug on the device preventing proper playback.